refactor: moved all interactions with LS to utility functions
